backend/requirements2.txt
Django==4.0.5
django-cors-headers==3.13.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
Pillow==9.2.0
PyJWT==2.4.0
